// backend/src/email/enums/email-type.enum.ts
export enum EmailType {
  USER_WELCOME = 'user_welcome',
  PASSWORD_RESET = 'password_reset',
  ACCOUNT_ACTIVATION = 'account_activation',
  LIMIT_WARNING = 'limit_warning',
  PAYMENT_FAILED = 'payment_failed',
  PAYMENT_SUCCEEDED_AFTER_FAILURE = 'payment_succeeded_after_failure',
  SUBSCRIPTION_CANCELED = 'subscription_canceled',
  GRACE_PERIOD_WARNING = 'grace_period_warning',
}